requirements.txt
openpyxl>=3.0.0
pyyaml>=5.0
reportlab>=3.5
matplotlib>=3.0
pillow>=8.0
imageio>=2.9
# Opcional, caso use gráficos de fluxos ou layouts avançados:
# graphviz>=0.20

# Tkinter já vem com o Python padrão (Windows/Mac). No Linux:
# sudo apt-get install python3-tk